+++ include/JSystem/JGadget/binary.h
@@ -10,9 +10,9 @@ struct TEBit {
 };
 
 struct TParse_header_block {
-	virtual ~TParse_header_block()                                             = 0; // _08
-	virtual bool parseHeader_next(const void**, unsigned long*, unsigned long) = 0; // _0C
-	virtual bool parseBlock_next(const void**, unsigned long*, unsigned long)  = 0; // _10
+	virtual ~TParse_header_block()                         = 0; // _08
+	virtual bool parseHeader_next(const void**, u32*, u32) = 0; // _0C
+	virtual bool parseBlock_next(const void**, u32*, u32)  = 0; // _10
 
 	bool parse_next(void const**, u32);
 };
